+++ plugin/src/components/OverviewPage.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import {initKialiListeners, kioskUrl, properties} from "../properties";
+import {initKialiListeners, kioskUrl, properties} from "../utils";
 import { consoleFetch } from "@openshift-console/dynamic-plugin-sdk";
 
 const OverviewPage = () => {
